extracted test endpoint and added instrumentation collector tests
