+++ test/lib/ex_debug_toolbar/phoenix_test.exs
@@ -1,25 +1,3 @@
-defmodule EndpointUsingExDebugToolbar do
-  use Phoenix.Endpoint, otp_app: :ex_debug_toolbar
-  require Logger
-  import Plug.Conn
-
-  plug Plug.RequestId
-
-  use ExDebugToolbar.Phoenix
-
-  plug :dummy_plug
-  def dummy_plug(conn, _) do
-    Logger.debug "log entry"
-    conn = conn
-    |> Plug.Conn.assign(:called?, true)
-    |> put_resp_content_type("text/html")
-    |> send_resp(200, "<html><body></body></html>")
-    if timeout = conn.assigns[:timeout] do
-      :timer.sleep timeout
-    end
-    conn
-  end
-end
 
 defmodule ExDebugToolbar.PhoenixTest do
   use ExUnit.Case, async: true
@@ -27,9 +5,10 @@ defmodule ExDebugToolbar.PhoenixTest do
   import Supervisor.Spec
   import ExDebugToolbar.Test.Support.RequestHelpers
   alias ExDebugToolbar.Data.Timeline
+  alias ExDebugToolbar.Fixtures.Endpoint
 
   setup_all do
-    children = [supervisor(EndpointUsingExDebugToolbar, [])]
+    children = [supervisor(Endpoint, [])]
     opts = [strategy: :one_for_one, name: ExDebugToolbarTest.Supervisor]
     Supervisor.start_link(children, opts)
     :ok
@@ -46,12 +25,12 @@ defmodule ExDebugToolbar.PhoenixTest do
   test "it tracks execution time of all following plugs in pipeline" do
     make_request timeout: 100
     assert {:ok, request} = get_request()
-    assert Timeline.duration(request.data.timeline) > 90 * 1000 # not sure why
+    assert Timeline.duration(request.data.timeline) > 70 * 1000 # not sure why
   end
 
   defp make_request(assigns \\ %{}) do
     conn(:get, "/")
     |> Map.put(:assigns, Map.new(assigns))
-    |> EndpointUsingExDebugToolbar.call(%{})
+    |> Endpoint.call(%{})
   end
 end
